+++ SoulKnife/SoulKnifeProficiencies.cs
@@ -11,7 +11,7 @@ public static class SoulKnifeProficiencies
         FeatureConfigurator.New("SoulKnifeProficiencies", Guids.SoulKnifeProficiencies)
             .SetDisplayName(Loc.Str("SK.Proficiencies.Name", "Soulknife Proficiencies"))
             .SetDescription(Loc.Str("SK.Proficiencies.Desc",
-                "A soulknife is proficient with all simple and martial weapons, with light and medium armor, and with shields (except tower shields)."))
+                "A soulknife is proficient with all simple and martial weapons, with light and medium armor, and with shields (except tower shields). She is always proficient with her mind blade, regardless of form."))
             .SetIcon(FeatureRefs.MartialWeaponProficiency.Reference.Get().Icon)
             .SetIsClassFeature()
             .AddFacts(